requirements_new.txt
cx-Freeze==6.3
cycler==0.10.0
importlib-metadata==2.1.3
kiwisolver==1.1.0
matplotlib==3.0.3
numpy==1.18.5
opencv-python==4.4.0.42
Pillow==7.2.0
pyparsing==2.4.7
python-dateutil==2.8.2
scipy==1.4.1
six==1.16.0
torch==1.1.0
torchvision==0.3.0
zipp==1.2.0
